Get real commit branch instead of the currently active one.
Previously it was assumed that the commit being processed is a member
of the currently active branch, which may not be the case.

This commit changes the behavior so that each time a commit is
processed - it's branches are got from the git tree. The first
found branch is used as a comit branch.

For this functionality to function properly - remote branches from
which the commits may be processed have to be checked out locally
beforehand using `git checkout -b <branch_name>`.

Finding the branch name for a specific commit is slow process which
can substantially increase the commit processing times.

Closes #32.